backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.3",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "uvicorn[standard]>=0.42.0",
]
